Создает модуль gallery
