+++ js/thumbnails.js
@@ -1,5 +1,5 @@
-import { getPhotos } from './data.js';
-import { showModal } from './modal.js';
+// import { getPhotos } from './data.js';
+import { renderGallery } from './gallery.js';
 
 const picturesContainer = document.querySelector('.pictures');
 
@@ -22,21 +22,6 @@ const createThumbnails = (getPhotos) => {
 
   picturesContainer.appendChild(picturesFragment);
   renderGallery(getPhotos);
-};
-
-const renderGallery = (getPhotos) => {
-  picturesContainer.addEventListener('click', (evt) => {
-    const thumbnail = evt.target.closest('[data-thumbnail-id]');
-
-    if (!thumbnail) {
-      return;
-    }
-    evt.preventDefault();
-
-    const thumbnailId = + thumbnail.dataset.thumbnailId;
-    const photoData = getPhotos.find(({ id }) => id === thumbnailId);
-    showModal(photoData);
-  });
 };
 
 export { createThumbnails };
